Refactor AboutActivity layout and apply edge-to-edge display
The root layout of `a_about.xml` has been changed from a `LinearLayout` to a `FrameLayout` and given the ID `layoutRoot`. The original `LinearLayout` is now a child of this `FrameLayout`.

In `AboutActivity.kt`, `UIUtils.setupEdgeToEdge1(window)` and `UIUtils.setupEdgeToEdge2(...)` have been called to implement an edge-to-edge display. The latter applies padding to the top and bottom of the `layoutRoot` view.